+++ HISTORY.rst
@@ -6,6 +6,7 @@ History
 -------------------
 
 * Card #552: LeadOrder events not being sent (ericof).
+* Improve logging of discarded events (ericof).
 
 2.0.19 (2017-08-07)
 -------------------